# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "DS ESTIMATION BACKEND DOCUMENTATION",
    "description": "Our story started in 2015 when there was a need of Construction database for helping bidders to bid accurately with reliable information's and estimates in the Rwanda construction industry. Continuous determination, hard work and passion have been constant themes throughout our history and present. As Rwanda Construction industry becomes increasingly important on the global stage, it's our believe that, with our own innovative mindset, we can make the construction industry a better one for national development. So we challenge ourselves to evolve, innovate and experiment to make work easier and faster for everyone as the leading local software brand in the Rwanda construction industry. We are currently the quantities and estimations software dealers which are made in Africa its self initially from Ghana and same technology has been customized to fit Rwanda construction industry and we are super proud to start the race with national development in minds. Today we are privileged to connect with thousands of Engineers and other construction related professionals who do daily estimates or quantities in the construction industry.",
    "version": "1.0.0",
    "contact": {
      "name": "DS ESTIMATION",
      "url": "http://dsestimation.com",
      "email": "dsestimation@gmail.com"
    }
  },
  "host": "dse-estimation-backend.herokuapp.com",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "AuthToken": {
      "type": "apiKey",
      "name": "auth-token",
      "in": "header",
      "description": "The token for authentication"
    }
  },
  "security": [
    {
      "AuthToken": []
    }
  ],
  "paths": {
...
    "/user/role/all": {
      "get": {
        "tags": ["User Role Routes"],
        "summary": "Get all roles from our database",
        "description": "Get all roles from our database",
        "operationId": "getAllRoles",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Roles Retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/user/role/create": {
      "post": {
        "tags": ["User Role Routes"],
        "summary": "Create a role",
        "description": "Create a role",
        "operationId": "createRole",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Role object that needs to be added to the database",
            "schema": {
              "$ref": "#/definitions/Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/role/{id}/update": {
      "patch": {
        "tags": ["User Role Routes"],
        "summary": "Update a role",
        "description": "Update a role",
        "operationId": "updateRole",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the role to update",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Role object that needs to be updated",
            "schema": {
              "$ref": "#/definitions/Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Role not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/role/{id}/delete": {
      "delete": {
        "tags": ["User Role Routes"],
        "summary": "Delete a role",
        "description": "Delete a role",
        "operationId": "deleteRole",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the role to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Role deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Role not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Role": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "permissions": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
  }
}